requirements.txt
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
